requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0
firebase-admin>=6.5.0
pyrebase4>=4.7.1
python-dotenv>=1.0.0
pandas>=2.2.0
Pillow>=10.3.0
pdf2image>=1.17.0
pymupdf>=1.23.0
PyPDF2>=3.0.0
httpx>=0.27.0
tenacity>=8.3.0
